scripts/image-generator/pyproject.toml
[project]
name = "image-generator"
version = "0.1.0"
description = "Generate images via OpenRouter API (Gemini)"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.0",
    "python-dotenv>=1.1.0",
]
